// src/dorc-web/src/apis/dorc-api/models/DatabaseAuditApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface DatabaseAuditApiModel
 */
export interface DatabaseAuditApiModel {
    /**
     * @type {number}
     * @memberof DatabaseAuditApiModel
     */
    Id?: number;
    /**
     * @type {number}
     * @memberof DatabaseAuditApiModel
     */
    DatabaseId?: number | null;
    /**
     * @type {string}
     * @memberof DatabaseAuditApiModel
     */
    DatabaseName?: string | null;
    /**
     * @type {number}
     * @memberof DatabaseAuditApiModel
     */
    RefDataAuditActionId?: number;
    /**
     * @type {string}
     * @memberof DatabaseAuditApiModel
     */
    Action?: string | null;
    /**
     * @type {string}
     * @memberof DatabaseAuditApiModel
     */
    Username?: string | null;
    /**
     * @type {string}
     * @memberof DatabaseAuditApiModel
     */
    Date?: string;
    /**
     * @type {string}
     * @memberof DatabaseAuditApiModel
     */
    FromValue?: string | null;
    /**
     * @type {string}
     * @memberof DatabaseAuditApiModel
     */
    ToValue?: string | null;
}