// src/dorc-web/src/apis/dorc-api/models/GetDatabaseAuditListResponseDto.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type {
    DatabaseAuditApiModel,
} from './';

/**
 * @export
 * @interface GetDatabaseAuditListResponseDto
 */
export interface GetDatabaseAuditListResponseDto {
    /**
     * @type {number}
     * @memberof GetDatabaseAuditListResponseDto
     */
    CurrentPage?: number;
    /**
     * @type {number}
     * @memberof GetDatabaseAuditListResponseDto
     */
    TotalItems?: number;
    /**
     * @type {number}
     * @memberof GetDatabaseAuditListResponseDto
     */
    TotalPages?: number;
    /**
     * @type {Array<DatabaseAuditApiModel>}
     * @memberof GetDatabaseAuditListResponseDto
     */
    Items?: Array<DatabaseAuditApiModel> | null;
}